% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group_output.R
\name{add_label}
\alias{add_label}
\title{Add label for posterior prediction plots}
\usage{
add_label(group_output)
}
\arguments{
\item{group_output}{extracted group output from \code{extract_group_output}}
}
\description{
Add label for posterior prediction plots
}
\keyword{internal}
