% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_assign_data.R
\name{aug_assign_data}
\alias{aug_assign_data}
\title{Assign data to template}
\usage{
aug_assign_data(model_template, ...)
}
\arguments{
\item{model_template}{output from aug_model}

\item{...}{data-masking name-value pairs}
}
\value{
updated named list with your data assigned to Stan model parameters
}
\description{
Assign data to template
}
\examples{
aug_model("linear") |> aug_assign_data(X = Loblolly$age, Y = Loblolly$height)
}
