% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_fit_model.R
\name{aug_fit_model}
\alias{aug_fit_model}
\title{Fit an Augmented Gaussian Function in Stan}
\usage{
aug_fit_model(transformed_data, ...)
}
\arguments{
\item{transformed_data}{data list object created using \code{transform_data()}}

\item{...}{arguments passed to rstan::sampling()}
}
\value{
large list object containing:
\itemize{
\item \code{stanfit}: Object of class stanfit containing output from the Stan model as returned by \code{sampling}
\item \code{diag}:
\item \code{samples}
\item \code{summary}
\item \code{waic}
\item \code{loo}
}
}
\description{
A short description...
}
\details{
The data must in a list object
\itemize{
\item \code{nSubj}: Number of subjects
\item \code{nStim}: Number of stimuli along the dimension
\item \code{xs}: Indices of stimuli on the x axis (0 represents the CS+)
}
}
\examples{
\dontrun{
demo_data |>
transform_data() |>
aug_fit_model(iter = 100, chains = 2)
}

}
\author{
Jessica Lee
}
