% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_models.R
\name{aug_model}
\alias{aug_model}
\title{Select data configuration template for aug supported model}
\usage{
aug_model(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
named list that matches Stan model parameters
}
\description{
Select data configuration template for aug supported model
}
\examples{
aug_model("linear")
}
