% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_run.R
\name{aug_run}
\alias{aug_run}
\title{Run a linear model in Stan}
\usage{
aug_run(model_template, ...)
}
\arguments{
\item{model_template}{model template generated by aug_model and updated by aug_assign_data}

\item{...}{additional arguments passed to rstan::sampling}
}
\value{
Stanfit model output
}
\description{
Run a linear model in Stan
}
\examples{
aug_model("linear") |>
  aug_assign_data(X = Loblolly$age,
                   Y = Loblolly$height,
                   N = nrow(Loblolly)) |>
  aug_run()
}
