% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group_output.R
\name{create_post_preds}
\alias{create_post_preds}
\title{Compile posterior predictions from model output}
\usage{
create_post_preds(group_output, nSamp = 50)
}
\arguments{
\item{group_output}{extracted group output from \code{extract_group_output}}

\item{nSamp}{...}
}
\description{
Compile posterior predictions from model output
}
\keyword{internal}
