% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{AugGauss Example Dataset}
\format{
\subsection{\code{demo_data}}{

A data frame with 630 rows and 4 columns:
\describe{
\item{subj}{Subject identifier}
\item{group}{Group identifier}
\item{x}{x}
\item{y}{y}
...
}
}
}
\usage{
demo_data
}
\description{
A subset of data from the XX
}
\keyword{datasets}
