% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdi.R
\name{Get_HDIs}
\alias{Get_HDIs}
\title{Calculate Highest Density Intervals and p(ROPE)}
\usage{
Get_HDIs(
  aug_model_output,
  ropeLow = c(-0.05, 0.1, 0.1, 70),
  ropeHigh = c(+0.05, 0.2, 0.2, 80),
  hdiLim = 0.95,
  propPost = 1,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{aug_model_output}{output from \code{aug_fit_model}.}

\item{ropeLow}{a vector of numbers specifying the lower limit of the ROPE for each of the 4 augmented Gaussian parameters in the order Mean, Height, Width-, Width+. Note: these limits are user-specified and should have good justification.}

\item{ropeHigh}{a vector of numbers specifying the lower limit of the ROPE for each of the 4 augmented Gaussian parameters in the order Mean, Height, Width-, Width+. Note: these limits are user-specified and should have good justification.}

\item{hdiLim}{a number specifying the probability mass used to compute the Highest Density Interval (HDI), defaults to .95.}

\item{propPost}{a number specifying the proportion of the posterior samples used to calculate p(direction) and p(ROPE), defaults to 1.}

\item{save_output}{a logical indicating whether to save output, defaults to FALSE.}

\item{path}{optional path to save output.}
}
\description{
This function calculates Highest Density Intervals (HDIs) and proportion of posterior samples falling within the Region of Practical Equivalence (ROPE) for the 4 parameters of the augmented Gaussian (Mean, Height, Width-, Width+).
}
