% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdi.R
\name{Get_HDIs_diff}
\alias{Get_HDIs_diff}
\title{Calculate Highest Density Intervals for differences between two groups}
\usage{
Get_HDIs_diff(
  aug_model_output,
  comparison = c("group1", "group2"),
  ropeLowDiffs = c(-0.05, -0.05, -0.05, -2.5),
  ropeHighDiffs = abs(c(-0.05, -0.05, -0.05, -2.5)),
  hdiLim = 0.95,
  propPost = 1,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{aug_model_output}{output from \code{aug_fit_model}}

\item{comparison}{vector of length 2 of names of groups to compare. These group names must match the names in the data.}

\item{ropeLowDiffs}{...}

\item{ropeHighDiffs}{...}

\item{propPost}{a number specifying the proportion of the posterior samples used to calculate p(direction) and p(ROPE), defaults to 1.}

\item{save_output}{a logical indicating whether to save output, defaults to FALSE.}

\item{path}{optional path to save output.}

\item{a}{number specifying the probability mass used to compute the Highest Density Interval (HDI), defaults to .95.}
}
\description{
#' @description
This function calculates Highest Density Intervals (HDIs) for differences in parameters between two groups.
}
