% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_preds.R
\name{Get_Posterior_Preds_by_group}
\alias{Get_Posterior_Preds_by_group}
\title{Get posterior predictions for each group}
\usage{
Get_Posterior_Preds_by_group(aug_model_output, group_nm, nSamp = 50)
}
\arguments{
\item{aug_model_output}{model output of \code{aug_fit_model}}

\item{group_nm}{name of the group}

\item{nSamp}{...}
}
\description{
Get posterior predictions for each group
}
\keyword{internal}
