% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_preds.R
\name{Plot_Posterior_Preds_by_group}
\alias{Plot_Posterior_Preds_by_group}
\title{Plot posterior predictions for a single group}
\usage{
Plot_Posterior_Preds_by_group(
  aug_model_output,
  group_name,
  nSamp = 50,
  include_preds = FALSE,
  labels = FALSE,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{aug_model_output}{model output from \code{aug_fit_model}.}

\item{group_name}{a character string specifying the name of the group.}

\item{nSamp}{an integer specifying the number of posterior samples to plot.}

\item{include_preds}{a logical indicating whether to include the predictions in the plot.}

\item{labels}{a logical indicating whether to display labels in the plot.}

\item{save_output}{a logical indicating whether to save output, defaults to FALSE.}

\item{path}{optional path to save output.}
}
\description{
Plot posterior predictions for a single group
}
