% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate augmented Gaussian gradients}
\usage{
simulate_data(
  nSubj,
  nGroups,
  dimVals = seq(-0.5, +0.5, 0.05),
  M = c(0, 0, 0.1),
  H = c(80, 75, 70),
  WM = c(0.2, 0.2, 0.2),
  WP = c(0.4, 0.4, 0.4),
  Noise = 2,
  save_data = FALSE,
  path = "output/"
)
}
\arguments{
\item{nSubj}{an integer specifying the number of subjects per group.}

\item{nGroups}{an integer specifying the number of experimental groups.}

\item{dimVals}{a vector of numbers specifying the dimension values. Must range between -0.5 and +0.5. Defaults to seq(-.5, +.5, .05).}

\item{M}{a vector of length nGroups specifying the Mean parameter for each group.}

\item{H}{a vector of length nGroups specifying the Height parameter for each group.}

\item{WM}{a vector of length nGroups specifying the Width- parameter for each group.}

\item{WP}{a vector of length nGroups specifying the Width+ parameter for each group.}

\item{Noise}{a number specifying the amount of noise (SD of Gaussian) added to simulate the data.}

\item{save_data}{a logical variable indicating whether to save data as .csv.}

\item{path}{path to save .csv output.}
}
\value{
tibble of simulated data
}
\description{
This function simulates augmented Gaussian gradients for groups of subjects.
}
\examples{
simulate_data
}
\author{
Jessica Lee
}
