% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_fit_model.R
\name{aug_fit_model}
\alias{aug_fit_model}
\title{Fit Augmented Gaussian Function}
\usage{
aug_fit_model(transformed_data, ...)
}
\arguments{
\item{transformed_data}{data list object created using \code{transform_data()}}

\item{...}{arguments passed to rstan::sampling()}
}
\value{
large list object containing:
\itemize{
\item \code{stanfit}: Object of class stanfit containing output from the Stan model as returned by \code{rstan::sampling()}.
\item \code{diag}: Diagnostics as returned by \code{rstan::get_sampler_params()}.
\item \code{samples}: posterior samples extracted from the stanfit object, as returned by \code{rstan::extract()}.
\item \code{summary}: summary statistics computed from the posterior samples, including mean, standard error, and quantiles at 2.5\%, 50\% (median), and 97.5\%.
\item \code{waic}: Widely Applicable Information Criterion (WAIC) computed on the log-likelihood as returned by \code{loo::waic()}.
\item \code{loo}: Efficient approximate lave-one-out cross-validation (LOO) computed on the log-likelihood as returned by \code{loo::loo()}.
}
}
\description{
Fits augmented Gaussian functions to individual gradients in a Stan model.
}
\details{
The data must in a list object.
\itemize{
\item \code{nSubj}: Number of subjects.
\item \code{nStim}: Number of stimuli along the dimension.
\item \code{xs}: Indices of stimuli on the x axis (0 represents the CS+).
}
}
\examples{
\dontrun{
demo_data |>
transform_data() |>
aug_fit_model(iter = 100, chains = 2)
}

}
\author{
Jessica Lee
}
