% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_run.R
\name{aug_fit_model}
\alias{aug_fit_model}
\title{Fit an Augmented Gaussian Bayesian Model}
\usage{
aug_fit_model(transformed_data, ...)
}
\arguments{
\item{transformed_data}{data list object created using \code{transform_data()}}

\item{...}{arguments passed to rstan::sampling()}
}
\value{
large list object containing ...
}
\description{
A short description...
}
\details{
The data must in a list object
\itemize{
\item \code{nSubj}
\item \code{nStim}
\item \code{xs}
}
}
\examples{
\dontrun{
demo_data |>
transform_data() |>
aug_fit_model(iter = 100, chains = 2)
}

}
\author{
Jessica Lee
}
