% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_preds.R
\name{extract_group_output}
\alias{extract_group_output}
\title{Extract output for posterior predictions}
\usage{
extract_group_output(aug_model_output, group_nm)
}
\arguments{
\item{aug_model_output}{model output of \code{aug_fit_model}.}

\item{group_nm}{a character string specifying the name of the group.}
}
\description{
Extract output for posterior predictions
}
\keyword{interal}
