% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdi.R
\name{Get_HDIs}
\alias{Get_HDIs}
\title{Calculate the Highest Density Intervals and p(Region of Practical Equivalence)}
\usage{
Get_HDIs(
  aug_model_output,
  ropeLow = c(-0.05, 0.1, 0.1, 70),
  ropeHigh = c(+0.05, 0.2, 0.2, 80),
  hdiLim = 0.95,
  propPost = 1,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{aug_model_output}{output from \code{aug_fit_model}}

\item{ropeLow}{...}

\item{ropeHigh}{...}

\item{hdiLim}{...}

\item{propPost}{...}

\item{save_output}{logical whether to save output}

\item{path}{path where output is saved}
}
\description{
Calculate the Highest Density Intervals and p(Region of Practical Equivalence)
}
