% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdi.R
\name{Get_HDIs_diff}
\alias{Get_HDIs_diff}
\title{Calculate the Highest Density Intervals and p(Region of Practical Equivalence) for HDIparams for differences between groups}
\usage{
Get_HDIs_diff(
  aug_model_output,
  comparison = c("group1", "group2"),
  ropeLowDiffs = c(-0.05, -0.05, -0.05, -2.5),
  ropeHighDiffs = abs(c(-0.05, -0.05, -0.05, -2.5)),
  hdiLim = 0.95,
  propPost = 1,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{aug_model_output}{output from \code{aug_fit_model}}

\item{comparison}{vector length of 2 of names of groups to compare}

\item{ropeLowDiffs}{...}

\item{ropeHighDiffs}{...}

\item{hdiLim}{...}

\item{propPost}{...}

\item{save_output}{logical whether to save output}

\item{path}{path where output is saved}
}
\description{
Calculate the Highest Density Intervals and p(Region of Practical Equivalence) for HDIparams for differences between groups
}
