% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group_output.R
\name{Plot_Posterior_Preds_by_group}
\alias{Plot_Posterior_Preds_by_group}
\title{Plot posterior predictions for a group}
\usage{
Plot_Posterior_Preds_by_group(
  group_output,
  nSamp = 50,
  include_preds = FALSE,
  labels = FALSE,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{group_output}{extracted group output from \code{extract_group_output}}

\item{nSamp}{...}

\item{include_preds}{logical include predictions or not}

\item{labels}{logical whether to display labels in plot}

\item{save_output}{logical whether to save output}

\item{path}{path to save output}
}
\description{
Plot posterior predictions for a group
}
