% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_preds.R
\name{Plot_Posterior_Preds_by_group}
\alias{Plot_Posterior_Preds_by_group}
\title{Plot posterior predictions for a group}
\usage{
Plot_Posterior_Preds_by_group(
  aug_model_output,
  group_name,
  nSamp = 50,
  include_preds = FALSE,
  labels = FALSE,
  save_output = FALSE,
  path = "output/"
)
}
\arguments{
\item{aug_model_output}{model output from \code{aug_fit_model}}

\item{group_name}{character string of one group}

\item{nSamp}{...}

\item{include_preds}{logical include predictions or not}

\item{labels}{logical whether to display labels in plot}

\item{save_output}{logical whether to save output}

\item{path}{path to save output}
}
\description{
Plot posterior predictions for a group
}
