% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulated_data.R
\name{plot_simulated_data}
\alias{plot_simulated_data}
\title{Plot simulated data nicely}
\usage{
plot_simulated_data(sim_data, dimVals = seq(-0.5, +0.5, 0.05))
}
\arguments{
\item{sim_data}{output from \code{simulate_data}.}

\item{dimVals}{a vector of numbers specifying the dimension values. Must range between -0.5 and +0.5. Defaults to seq(-.5, +.5, .05).}
}
\description{
Plot simulated data nicely
}
