% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate augmented Gaussian gradients}
\usage{
simulate_data(
  nSubj,
  nGroups,
  dimVals = seq(-0.5, +0.5, 0.05),
  M = c(0, 0, 0.1),
  H = c(80, 75, 70),
  WM = c(0.2, 0.2, 0.2),
  WP = c(0.4, 0.4, 0.4),
  Noise = 2,
  save_data = FALSE,
  path = "output/"
)
}
\arguments{
\item{nSubj}{number of subjects}

\item{nGroups}{number of experimental groups}

\item{dimVals}{dimension values}

\item{M}{...}

\item{H}{...}

\item{WM}{...}

\item{WP}{...}

\item{Noise}{...}

\item{save_data}{logical, save data as .csv}

\item{path}{path to save .csv output}
}
\value{
tibble of simulated data
}
\description{
This function simulates augmented Gaussian gradients for groups of subjects
}
\examples{
simulate_data
}
\author{
Jessica Lee
}
