% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform tabular data in Stan-friendly list data}
\usage{
transform_data(sim_data, dimVals = seq(-0.5, +0.5, 0.05), groupNames = NULL)
}
\arguments{
\item{sim_data}{output from \code{simulate_data}}

\item{dimVals}{a vector of numbers specifying the dimension values. Must range between -0.5 and +0.5. Defaults to seq(-.5, +.5, .05).}

\item{groupNames}{optional vector of group names. If NULL, group names will be read from data.}
}
\value{
list object of length(nGroups) containing datalists for Stan.
}
\description{
Transform tabular data in Stan-friendly list data
}
